+++ src/kiosk/internal_info/router.py
@@ -15,6 +15,7 @@ from src.kiosk.internal_info.schemas import (
     ResponseModel,
 )
 from src.database import get_db
+from src.enums import DocumentType
 from src.upload import delete_file
 from src.logger import app_logger
 from src.auth import get_auth_user
@@ -156,7 +157,8 @@ def delete_internal_info(item_id: int, db: Session = Depends(get_db)) -> Respons
         if item.thumbnail_path:
             delete_file(item.thumbnail_path)
         for doc in item.documents:
-            delete_file(doc.file_path)
+            if doc.type != DocumentType.YOUTUBE:
+                delete_file(doc.file_path)
 
         db.delete(item)
         db.commit()
@@ -232,7 +234,8 @@ def delete_document(item_id: int, doc_id: int, db: Session = Depends(get_db)) ->
         if not doc:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Document not found.")
 
-        delete_file(doc.file_path)
+        if doc.type != DocumentType.YOUTUBE:
+            delete_file(doc.file_path)
         db.delete(doc)
         db.commit()
         return ResponseModel()
